@@ -1,13 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Vintagestory.API.Datastructures;
+﻿using Vintagestory.API.Datastructures;
 
-namespace AutoLootReforged.Code.Custom
-{
-    public class HarvestableListener : TreeModifiedListener
-    {
-    }
-}
+namespace AutoLootReforged.Code.Custom;
+
+/// <summary>
+/// An override of TreeModifiedListener so we can distinguish
+/// </summary>
+public class HarvestableListener : TreeModifiedListener;
